update and added new files

+++ src/main/java/com/example/demo/service/UserService.java
@@ -5,11 +5,13 @@ import com.example.demo.model.User;
 import java.util.List;
 import java.util.Optional;
 
+
 public interface UserService
 {
     User createUser(User user);
-    Optional<User> getUserById(Integer id);
+    Optional<User> getUserById(Long id);
     List<User> getAllUsers();
     User updateUser(User user);
     void deleteUser(Integer id);
+
 }
